-Added a few images from wiiflow-mod, and actually use them here. -Update freetype, libpng, and remove metaphrasis -Fix flip sounds (The correct way!) -Add ability to set the code dump reload timeout by ini, default was 5 seconds, under [DEBUG], dsi_timeout, wifi gecko settings are also moved to the [DEBUG] domain. -Async network is automatically initialized once the menu is up now is wifi gecko is enabled -Remove default values from calls to CConfig -Fix defines.h -Refine fonts and sounds loading (There ARE changes in the theme.ini lines to be documented) -Fix game removal, the correct way. -Add new homebrew view, in about 250 lines less code than the version fix94 committed at wiiflow-mod in an attempt to imitate my preview.  It can be disabled by ini but is now the default. -Got some remnants of mod code in there still till i get it all weeded through eventually... and add my own dml ofc. -Left a few testicals laying in there for the high school script kiddies -Clean up a shetload of code, ALOT still to go, and alot of changes to merge from 1 dirty source to this now.
--Issues known to still exist: I get a freeze on the wait screen during mass b+reload, a lag in coverflow when switching from game to game during banner sound extraction, nand emu does not stat and read cache on first boot still, still cant hit channel view with no sd/hdd inserted.



git-svn-id: https://wiiflow-advanced.googlecode.com/svn/trunk@418 43c7226f-5572-e80b-0314-b0280708705f

+++ source/loader/wbfs_ext.c
@@ -148,13 +148,9 @@ void WBFS_Ext_ClosePart(wbfs_t* part)
 
 s32 WBFS_Ext_RemoveGame(u8 *discid, char *gamepath)
 {
-	//split_create(&split, gamepath, 0, 0, true);
-	//split_close(&split);
-
 	DIR *dir_iter;
 	struct dirent *ent;
 	char file[MAX_FAT_PATH];
-
 	char folder[MAX_FAT_PATH];
 	STRCOPY(folder, gamepath);
 	char *p = strrchr(folder, '/');
@@ -167,28 +163,38 @@ s32 WBFS_Ext_RemoveGame(u8 *discid, char *gamepath)
 		if (ent->d_name[0] == '.') continue;
 
 		snprintf(file, sizeof(file), "%s/%s", folder, ent->d_name);
-		remove(file);
-		break;
+		if(discid && strstr(file, (char *)discid) != NULL)
+			remove(file);
 	}
+
 	closedir(dir_iter);
-	unlink(folder);
+
+	if(strlen(folder) > 11)
+		unlink(folder);
+
 	return 0;
 }
 
 s32 WBFS_Ext_AddGame(progress_callback_t spinner, void *spinner_data)
 {
-	struct discHdr header ATTRIBUTE_ALIGN(32);
-
 	char folder[MAX_FAT_PATH];
 	bzero(folder, MAX_FAT_PATH);
 
+	snprintf(folder, sizeof(folder), "%s%s", wbfs_fs_drive, wbfs_ext_dir);
+	makedir((char *)folder);
+
+	struct discHdr header ATTRIBUTE_ALIGN(32);
+	Disc_ReadHeader(&header);
+
+	char *illegal = "\"*/:<>?\\|", *cp, *cleantitle;
+	asprintf(&cleantitle, header.title);
+	for (cp = strpbrk(cleantitle, illegal); cp; cp = strpbrk(cp, illegal))
+		if(cp) *cp = '_';
+	
 	char gamepath[MAX_FAT_PATH];
 	bzero(gamepath, MAX_FAT_PATH);
-	
-	Disc_ReadHeader(&header);
-	snprintf(folder, sizeof(folder), "%s%s/%s [%s]", wbfs_fs_drive, wbfs_ext_dir, header.title, header.id);
-	makedir((char *)folder);
-	snprintf(gamepath, sizeof(gamepath), "%s/%s.wbfs", folder, header.id);
+	snprintf(gamepath, sizeof(gamepath), "%s/%s [%s].wbfs", folder, cleantitle, header.id);
+	free(cleantitle);
 
 	u64 size = (u64)143432*2*0x8000ULL;
 	u32 n_sector = size / 512;
@@ -222,7 +228,7 @@ s32 WBFS_Ext_AddGame(progress_callback_t spinner, void *spinner_data)
 
 	WBFS_Ext_ClosePart(part);
 	
-	if(ret < 0) WBFS_Ext_RemoveGame(NULL, gamepath);
+	if(ret < 0) WBFS_Ext_RemoveGame(header.id, gamepath);
 
 	return ret < 0 ? ret : 0;
 }
